# webcore_appcore_starter_4_17/backend/control_plane/openapi.v1.yaml
openapi: 3.0.3
info:
  title: Control Plane IAM API
  version: 1.0.0
  description: |
    Multi-tenant IAM foundation: Tenant/Project/Environment + RBAC + OIDC auth + immutable AuditLog.
    
    Fail-Closed: default deny. No token or insufficient permission => 401/403.
    Multi-tenancy: every record scoped by tenant_id; cross-tenant access blocked by design.
    Audit: every mutating API writes append-only audit record (no updates/deletes).
    Deterministic permissions: same token + same state => same decision.
servers:
  - url: https://api.example.com/api/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Local development server
security:
  - BearerAuth: []
paths:
  /iam/tenants:
    get:
      summary: List tenants
      description: List all tenants (admin only, tenant-scoped)
      operationId: listTenants
      tags:
        - Tenants
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenants:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tenant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      summary: Create tenant
      description: Create a new tenant
      operationId: createTenant
      tags:
        - Tenants
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenant:
                    $ref: '#/components/schemas/Tenant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /iam/tenants/{id}:
    get:
      summary: Get tenant by ID
      description: Get tenant details (tenant-scoped, cross-tenant access blocked)
      operationId: getTenant
      tags:
        - Tenants
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenant:
                    $ref: '#/components/schemas/Tenant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update tenant
      description: Update tenant (tenant-scoped, cross-tenant access blocked)
      operationId: updateTenant
      tags:
        - Tenants
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenantRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenant:
                    $ref: '#/components/schemas/Tenant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC JWT token (verified via issuer/JWKS)
  schemas:
    Tenant:
      type: object
      required:
        - id
        - name
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [active, suspended, deleted]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
    CreateTenantRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        metadata:
          type: object
    UpdateTenantRequest:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, suspended]
        metadata:
          type: object
...
  responses:
    Unauthorized:
      description: Unauthorized (no token or invalid token)
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Unauthorized
              message:
                type: string
    Forbidden:
      description: Forbidden (insufficient permission or cross-tenant access)
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Forbidden
              message:
                type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Not found